chore: clean up format / lint / clippy drift to unblock check gate
Why
Running the full `bun run check` after BROWSE-VIRT closeout surfaced
36 pre-existing eslint errors (mostly @typescript-eslint/require-await,
no-unnecessary-type-assertion, react-hooks/set-state-in-effect, plus
the unused-eslint-disable from my own viewport-mount hook), 35 files
with Prettier drift, two Rust clippy errors (derivable Default on
OgImageFetchMode + unnecessary-cast i64→i64 in og_images test seed),
and one dead struct (BundlePathPayload was a leftover from the S3
removal — no callers since the cloud-backup commands were removed).
Without these fixes the check gate can't even reach the coverage /
e2e / mutation stages.

What
- Prettier auto-format pass over 35 files (mostly settings tests +
  hook test fixtures the previous agents had not run through
  `bun run format`). Mechanical reformat — no behavioural change.
- ESLint cleanup:
  * `use-route-history-nav.test.tsx`: drop `async` from `test(...)`
    callbacks + the inner `act(...)` arrows since none of them
    actually awaited anything (the lint rules
    `@typescript-eslint/require-await` + `@typescript-eslint/await-thenable`
    now reject this pattern).
  * `use-route-history-nav.ts`: add a targeted
    `eslint-disable-next-line react-hooks/set-state-in-effect` with a
    why-comment for the one legitimate setState-in-effect (the hook
    integrates router navigation events into the back/forward counter
    — there is no derived-state path that captures both Push +1 and
    Pop -1 honestly).
  * `link-previews-section.test.tsx` / `paper-form-primitives.test.tsx`:
    restore the `<HTMLButtonElement>` / `<HTMLInputElement>` generic
    on `screen.getByTestId(...)` callsites that need `.disabled` —
    `eslint --fix` had stripped them because the lint rule couldn't
    see the downstream `.disabled` access, but TypeScript still
    requires the narrowing.
  * `use-viewport-mount.test.tsx`: defer the onState side-effect to
    `useEffect` so the render itself stays pure (the new
    `react-hooks/access-refs-during-render` rule flagged the inline
    `onState(state)` call).
  * Drop the now-unused `eslint-disable-next-line` directive in
    `use-viewport-mount.ts` after switching the effect to the
    functional setter form.
  * Drop the unused `// eslint-disable-next-line no-console` in the
    BROWSE-VIRT spike test.
- Rust clippy:
  * Replace `impl Default for OgImageFetchMode { fn default() -> Self
    { Self::Background } }` with `#[derive(Default)]` + `#[default]`
    on the `Background` variant.
  * Drop the `(id * 1000) as i64` cast in
    `archive/history/og_images.rs` test seed (the binding is already
    `i64`).
  * Delete the unused `BundlePathPayload` struct in
    `src-tauri/src/dev_ipc_bridge/payloads.rs`. It was a leftover
    from the S3 verifier removal.
- Coverage: add 3 cases to `src/lib/backend-client.test.ts` for
  `exportAppData` / `previewAppDataImport` / `applyAppDataImport` so
  `src/lib/backend-client/migration.ts` is exercised end-to-end
  (was 25 % stmts → 100 % stmts).
- Spike test: bump per-scenario timeout to 60 s on the largest two
  BROWSE-VIRT spike scenarios (5 000 rows × cards mode) because the
  v8 coverage instrumentation slows them past the default 5 s
  budget; smaller scenarios keep the default.

How
- All product-code changes are mechanical hygiene; the
  `BundlePathPayload` removal is the only behavioural change and is
  safe because no caller has constructed it since the S3 cloud-backup
  removal commit (`9e0a266`).
- The `setState-in-effect` disable carries a why-comment so future
  readers know the lint rule was suppressed deliberately, not
  papered over.

Verification
- `bun run check:js` — format + lint + i18n + typecheck + 1947 unit
  tests all green.
- `bun run check:desktop-contract` — 100 % coverage on the bridge
  surface.
- `bun run check:rust` — fmt + clippy + 686 cargo tests across
  vault-core (570) / vault-worker (64) / pathkeep-desktop (46) /
  vault-platform (5) / browser-history-parser (1) all green.
- `bun run check:supply-chain` — audit + cargo-deny clean (existing
  RUSTSEC waivers documented in the audit output).
- `bun run check:platform` — 10/10 unit tests pass.
- `bun run release:check` — release configuration verified.

Residual
- `bun run coverage:js` global thresholds (99 % stmts / 98 %
  branches) still failing at 98.85 / 97.72. This is a *pre-existing*
  residual already tracked by `WORK-V03-COVERAGE-RESIDUAL`
  (BACKLOG.md). The migration.ts addition above pushed migration
  client from 25 → 100, but the remaining gap is concentrated in
  `src/app/shell.tsx`, `src/pages/dashboard/index.tsx`, several
  `src/components/explorer-paper/*` defensive guards
  (toLocaleString try/catch, SSR-only Number.isNaN fallbacks, ref
  null guards) and the per-route index.tsx files — exactly the set
  enumerated in the backlog entry's residual list. The hygiene
  pass above does not regress coverage; it improves it slightly
  (98.84 → 98.85 stmts).
- `bun run test:e2e` / `bun run test:e2e:desktop-bridge:truth` —
  not runnable in this headless Linux dev session (need Playwright
  browsers + display). User direction "A" (run heavy gate
  together) takes effect now, and per per-AGENTS.md `bun run
  check:deep` covers full mutation; those will run during the
  user's real-desktop verification pass.

+++ src-tauri/crates/vault-core/src/archive/history/og_images.rs
@@ -763,12 +763,7 @@ mod tests {
     fn list_urls_for_prefetch_honors_the_limit() {
         let connection = open_test_archive();
         for id in 1..=5 {
-            seed_url(
-                &connection,
-                id,
-                &format!("https://example.com/page/{id}"),
-                (id * 1000) as i64,
-            );
+            seed_url(&connection, id, &format!("https://example.com/page/{id}"), id * 1000);
         }
 
         let two = list_urls_for_prefetch(&connection, 2).unwrap();
